pass useReactResponsiveImageData in react-responsive-image
