--- packages/react-hooks/use-responsive-url/.gitignore
@@ -1,3 +1,0 @@
-.idea
-.cache
-node_modules
