--- packages/react-hooks/use-responsive-url/.npmignore
@@ -1,4 +1,0 @@
-.idea
-.cache
-node_modules
-src
