--- packages/react-hooks/use-responsive-url/README.md
@@ -1,21 +1,0 @@
-# @wbe/react-use-responsive-url
-
-Select responsive url in terms of window width
-
-![](https://img.shields.io/npm/v/@wbe/react-use-responsive-url/latest.svg)
-![](https://img.shields.io/bundlephobia/minzip/@wbe/react-use-responsive-url.svg)
-![](https://img.shields.io/david/willybrauner/libraries.svg?path=packages%2Freact-components%2Freact-use-responsive-url)
-![](https://img.shields.io/npm/dt/@wbe/react-use-responsive-url.svg)
-![](https://img.shields.io/npm/l/@wbe/react-use-responsive-url.svg)
-
-## Installation
-
-```shell script
-$ npm install @wbe/react-use-responsive-url
-```
-
-## How to use
-
-```js
-import { reactUseResponsiveUrl } from "@wbe/react-use-responsive-url;";
-```
