--- packages/react-hooks/use-responsive-url/package.json
@@ -1,36 +1,0 @@
-{
-  "name": "@wbe/use-responsive-url",
-  "version": "0.0.1",
-  "description": "Select responsive url in terms of window width",
-  "author": "Willy Brauner",
-  "license": "MIT",
-  "main": "dist/index",
-  "types": "dist/index",
-  "private": true,
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/willybrauner/libraries.git"
-  },
-  "scripts": {
-    "build": "npm run clean && npm run compile",
-    "clean": "rm -rf ./dist && rm -rf tsconfig.tsbuildinfo",
-    "compile": "tsc -b tsconfig.json"
-  },
-  "dependencies": {
-    "@wbe/react-responsive-image": "^0.0.1",
-    "@wbe/use-window-size": "^0.0.23",
-    "react": "^16.12.0",
-    "react-dom": "^16.12.0"
-  },
-  "devDependencies": {
-    "@types/react": "^16.9.16",
-    "@types/react-dom": "^16.9.4",
-    "typescript": "^3.7.3"
-  },
-  "keywords": [
-    "react"
-  ],
-  "publishConfig": {
-    "access": "public"
-  }
-}
