+++ packages/storybook/package.json
@@ -17,7 +17,6 @@
     "@wbe/react-responsive-image": "^0.0.1",
     "@wbe/use-bounding-client-rect": "^0.0.23",
     "@wbe/use-is-in-viewport": "^0.0.23",
-    "@wbe/use-responsive-url": "^0.0.1",
     "@wbe/use-window-size": "^0.0.23"
   },
   "devDependencies": {
